+++ app/javascript/applications/HomebrewsApp/pages/Content/Daggerheart/Items.jsx
@@ -87,7 +87,7 @@ export const DaggerheartItems = () => {
   const [open, setOpen] = createSignal(false);
 
   const [appState] = useAppState();
-  const [{ renderAlert, renderNotice }] = useAppAlert();
+  const [{ renderAlert, renderNotice, renderAlerts }] = useAppAlert();
   const [locale] = useAppLocale();
   const { Modal, openModal, closeModal } = createModal();
 
@@ -157,7 +157,7 @@ export const DaggerheartItems = () => {
         setConsume([]);
         closeModal();
       });
-    }
+    } else renderAlerts(result.errors_list);
   }
 
   const updateItem = async () => {
@@ -195,7 +195,7 @@ export const DaggerheartItems = () => {
           closeModal();
         });
       }
-    }
+    } else renderAlerts(result.errors_list);
   }
 
   const removeItem = async (itemId) => {
@@ -203,7 +203,7 @@ export const DaggerheartItems = () => {
 
     if (result.errors_list === undefined) {
       setItems(items().filter(({ id }) => id !== itemId ));
-    }
+    } else renderAlerts(result.errors_list);
   }
 
   const copyItem = async (itemId) => {
@@ -212,7 +212,7 @@ export const DaggerheartItems = () => {
     if (item.errors_list === undefined) {
       setItems([item.item].concat(items()));
       renderNotice(TRANSLATION[locale()].copyCompleted);
-    }
+    } else renderAlerts(result.errors_list);
   }
 
   const addToBook = async () => {
@@ -224,7 +224,7 @@ export const DaggerheartItems = () => {
         setSelectedIds([]);
       });
       renderNotice(TRANSLATION[locale()].added)
-    }
+    } else renderAlerts(result.errors_list);
   }
 
   return (